Add files via upload
def traverse(tree, init): BFS traversal function. This func accepts two inputs: the tree to traverse and the initial state. We already provided an example of how to run this func in the as1_test.py. Your task is to modify it to run with the loopy path. So far, if you try to traverse the tree_2 without a valid modification, your program crashes. 
def pathfinder(tree, init, goal): still BFS traversal function. However, this func accepts three inputs: the tree, the initial state, and the goal state. Your task is to find the solution from the initial state to the goal state even we have the loopy path in the tree.