Add files via upload
In this assignment, we will consider the problem of finding the shortest path for a robot vacuum on an 8x8 grid. The detailed instructions are included in the PA2_instruction.pdf file and python files that will guide you through the assignment.

The code for this assignment includes the following files:

PA2_search.py: you will edit this file which contains methods for A* an UCS algorithms.
PA2_test.py: this file is used to visualize the results and test the A* and UCS algorithms.
Files to Edit and Submit: You will fill in portions of PA2_search.py only during the assignment, and submit it. You should submit this file with your code and comments. Please do not change the other files in this distribution or submit any of our original files other than this file.
In PA2_search.py file, there are ###### START OF YOUR CODE HERE/#### END OF YOUR CODE#### tags denoting the start and end of code sections you should fill out. Take care to not delete or modify these tags. Otherwise, your assignment may not be properly graded.

Complete UCsearch() and aStarSearch() functions in the PA2_search.py file.

Although there are multiple solutions with A* and UCS, make sure that you achieve the same cost for these methods.

To test your implementation, run your code with: python PA2_test.py