Add files via upload
Tasks: two tasks need to be fulfilled:

There are four (4) logical errors, which are intentionally put inside the file: as3_mnx.py. You have to find ALL the errors and fix them. Otherwise, your code will not work correctly. For example, the result of the best terminal value in the root should be '3' instead of '6'. 
Implement the additional code (also in the file: as3_mnx.py) to return the traversed_solution in the form of an array. For example: ['A', 'X', 'Y'], with A is the root node, X is the child of A, Y is the leaf node containing the best terminal value (which is 3). Apparently, Y the child of X.